requirements.txt
pymysql
pymongo
python-dotenv



"""Настройте интерактивное меню, где пользователь может:
Выполнить поиск по ключевому слову.
Выполнить поиск по жанру и диапазону годов.
Посмотреть популярные или последние запросы.
Выйти из приложения.
"""